requirements/base.txt
Django==4.2
environs[django]==9.5.0
django-phonenumber-field[phonenumbers]==7.1.0
django-redis==5.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
# djangorestframework_simplejwt==5.2.2
psycopg2-binary==2.9.6
